+++ redasm/disassembler/listing/listingcursor.h
@@ -9,7 +9,7 @@ namespace REDasm {
 class ListingCursor
 {
     public:
-        typedef std::pair<u64, u64> Position;       // [Line, Column]
+        typedef std::pair<size_t, size_t> Position;       // [Line, Column]
         typedef std::stack<Position> PositionStack;
 
     public:
@@ -28,22 +28,22 @@ class ListingCursor
         const ListingCursor::Position& currentSelection() const;
         const ListingCursor::Position& startSelection() const;
         const ListingCursor::Position& endSelection() const;
-        u64 currentLine() const;
-        u64 currentColumn() const;
-        u64 selectionLine() const;
-        u64 selectionColumn() const;
-        bool isLineSelected(u64 line) const;
+        size_t currentLine() const;
+        size_t currentColumn() const;
+        size_t selectionLine() const;
+        size_t selectionColumn() const;
+        bool isLineSelected(size_t line) const;
         bool hasSelection() const;
         bool canGoBack() const;
         bool canGoForward() const;
-        void set(u64 line, u64 column = 0);
-        void moveTo(u64 line, u64 column = 0);
-        void select(u64 line, u64 column = 0);
+        void set(size_t line, size_t column = 0);
+        void moveTo(size_t line, size_t column = 0);
+        void select(size_t line, size_t column = 0);
         void goBack();
         void goForward();
 
     private:
-        void moveTo(u64 line, u64 column, bool save);
+        void moveTo(size_t line, size_t column, bool save);
 
     private:
         Position m_position, m_selection;
